// src/app/models/res-menu.model.ts
export interface ResMenu {
    find: any;
        id: string;
        name: string;
        category_menu_id: string;
        meal_time_id: string;
        min_quantity: string;
        stock_quantity: string;
        order_restriction_id: string;
        description: string;
        price: string;
        restart_stock: string;
        image: string;
        production_minutes: string;
        cooking_minutes: string;
        status: string;
        created_at: string;
        updated_at: string;
}